WTF is this?
Coming from someone who doesn't accept contributions that don't follow
[a particular style][cs], this is a terrible commit with a terrible
message.

Why was this change made?  What are the differences?  Trying to diff
this pile of crap answers neither question.

Plus, I'm only coming across this in the first place, because it
(might?) be the cause of my workaround for disabling the `:Gdelete`
command in order to make `:Gd` unambiguously refer to `:Gdiff`.

Maybe, just maybe, someday I'll use a `tpope`-authored Vim plugin
without having to fork the damn thing in order to delete the terrible
default mappings and/or configuration.

[cs]: https://commit.style